agent: refactor State
This allows for the State to be shared with the ssh agent.
