+++ src/bin/rbw-agent/actions.rs
@@ -77,7 +77,7 @@ pub async fn register(
 
 pub async fn login(
     sock: &mut crate::sock::Sock,
-    state: std::sync::Arc<tokio::sync::Mutex<crate::agent::State>>,
+    state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
     environment: &rbw::protocol::Environment,
 ) -> anyhow::Result<()> {
     let db = load_db().await.unwrap_or_else(|_| rbw::db::Db::new());
@@ -302,7 +302,7 @@ async fn two_factor(
 }
 
 async fn login_success(
-    state: std::sync::Arc<tokio::sync::Mutex<crate::agent::State>>,
+    state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
     access_token: String,
     refresh_token: String,
     kdf: rbw::api::KdfType,
@@ -358,7 +358,7 @@ async fn login_success(
 
 pub async fn unlock(
     sock: &mut crate::sock::Sock,
-    state: std::sync::Arc<tokio::sync::Mutex<crate::agent::State>>,
+    state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
     environment: &rbw::protocol::Environment,
 ) -> anyhow::Result<()> {
     if state.lock().await.needs_unlock() {
@@ -447,7 +447,7 @@ pub async fn unlock(
 }
 
 async fn unlock_success(
-    state: std::sync::Arc<tokio::sync::Mutex<crate::agent::State>>,
+    state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
     keys: rbw::locked::Keys,
     org_keys: std::collections::HashMap<String, rbw::locked::Keys>,
 ) -> anyhow::Result<()> {
@@ -459,7 +459,7 @@ async fn unlock_success(
 
 pub async fn lock(
     sock: &mut crate::sock::Sock,
-    state: std::sync::Arc<tokio::sync::Mutex<crate::agent::State>>,
+    state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
 ) -> anyhow::Result<()> {
     state.lock().await.clear();
 
@@ -470,7 +470,7 @@ pub async fn lock(
 
 pub async fn check_lock(
     sock: &mut crate::sock::Sock,
-    state: std::sync::Arc<tokio::sync::Mutex<crate::agent::State>>,
+    state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
 ) -> anyhow::Result<()> {
     if state.lock().await.needs_unlock() {
         return Err(anyhow::anyhow!("agent is locked"));
@@ -483,7 +483,7 @@ pub async fn check_lock(
 
 pub async fn sync(
     sock: Option<&mut crate::sock::Sock>,
-    state: std::sync::Arc<tokio::sync::Mutex<crate::agent::State>>,
+    state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
 ) -> anyhow::Result<()> {
     let mut db = load_db().await?;
 
@@ -525,7 +525,7 @@ pub async fn sync(
 
 pub async fn decrypt(
     sock: &mut crate::sock::Sock,
-    state: std::sync::Arc<tokio::sync::Mutex<crate::agent::State>>,
+    state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
     cipherstring: &str,
     entry_key: Option<&str>,
     org_id: Option<&str>,
@@ -564,7 +564,7 @@ pub async fn decrypt(
 
 pub async fn encrypt(
     sock: &mut crate::sock::Sock,
-    state: std::sync::Arc<tokio::sync::Mutex<crate::agent::State>>,
+    state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
     plaintext: &str,
     org_id: Option<&str>,
 ) -> anyhow::Result<()> {
@@ -588,7 +588,7 @@ pub async fn encrypt(
 #[cfg(feature = "clipboard")]
 pub async fn clipboard_store(
     sock: &mut crate::sock::Sock,
-    state: std::sync::Arc<tokio::sync::Mutex<crate::agent::State>>,
+    state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
     text: &str,
 ) -> anyhow::Result<()> {
     let mut state = state.lock().await;
@@ -606,7 +606,7 @@ pub async fn clipboard_store(
 #[cfg(not(feature = "clipboard"))]
 pub async fn clipboard_store(
     sock: &mut crate::sock::Sock,
-    _state: std::sync::Arc<tokio::sync::Mutex<crate::agent::State>>,
+    _state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
     _text: &str,
 ) -> anyhow::Result<()> {
     sock.send(&rbw::protocol::Response::Error {
@@ -693,7 +693,7 @@ async fn config_pinentry() -> anyhow::Result<String> {
 }
 
 pub async fn subscribe_to_notifications(
-    state: std::sync::Arc<tokio::sync::Mutex<crate::agent::State>>,
+    state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
 ) -> anyhow::Result<()> {
     if state.lock().await.notifications_handler.is_connected() {
         return Ok(());
