+++ src/bin/rbw-agent/agent.rs
@@ -1,83 +1,23 @@
 use anyhow::Context as _;
 use futures_util::StreamExt as _;
 
-pub struct State {
-    pub priv_key: Option<rbw::locked::Keys>,
-    pub org_keys:
-        Option<std::collections::HashMap<String, rbw::locked::Keys>>,
-    pub timeout: crate::timeout::Timeout,
-    pub timeout_duration: std::time::Duration,
-    pub sync_timeout: crate::timeout::Timeout,
-    pub sync_timeout_duration: std::time::Duration,
-    pub notifications_handler: crate::notifications::Handler,
-    #[cfg(feature = "clipboard")]
-    pub clipboard: Option<arboard::Clipboard>,
-}
-
-impl State {
-    pub fn key(&self, org_id: Option<&str>) -> Option<&rbw::locked::Keys> {
-        org_id.map_or(self.priv_key.as_ref(), |id| {
-            self.org_keys.as_ref().and_then(|h| h.get(id))
-        })
-    }
-
-    pub fn needs_unlock(&self) -> bool {
-        self.priv_key.is_none() || self.org_keys.is_none()
-    }
-
-    pub fn set_timeout(&self) {
-        self.timeout.set(self.timeout_duration);
-    }
-
-    pub fn clear(&mut self) {
-        self.priv_key = None;
-        self.org_keys = None;
-        self.timeout.clear();
-    }
-
-    pub fn set_sync_timeout(&self) {
-        self.sync_timeout.set(self.sync_timeout_duration);
-    }
-}
-
 pub struct Agent {
     timer_r: tokio::sync::mpsc::UnboundedReceiver<()>,
     sync_timer_r: tokio::sync::mpsc::UnboundedReceiver<()>,
-    state: std::sync::Arc<tokio::sync::Mutex<State>>,
+    state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
 }
 
 impl Agent {
-    pub fn new() -> anyhow::Result<Self> {
-        let config = rbw::config::Config::load()?;
-        let timeout_duration =
-            std::time::Duration::from_secs(config.lock_timeout);
-        let sync_timeout_duration =
-            std::time::Duration::from_secs(config.sync_interval);
-        let (timeout, timer_r) = crate::timeout::Timeout::new();
-        let (sync_timeout, sync_timer_r) = crate::timeout::Timeout::new();
-        if sync_timeout_duration > std::time::Duration::ZERO {
-            sync_timeout.set(sync_timeout_duration);
-        }
-        let notifications_handler = crate::notifications::Handler::new();
-        Ok(Self {
+    pub fn new(
+        timer_r: tokio::sync::mpsc::UnboundedReceiver<()>,
+        sync_timer_r: tokio::sync::mpsc::UnboundedReceiver<()>,
+        state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
+    ) -> Self {
+        Self {
             timer_r,
             sync_timer_r,
-            state: std::sync::Arc::new(tokio::sync::Mutex::new(State {
-                priv_key: None,
-                org_keys: None,
-                timeout,
-                timeout_duration,
-                sync_timeout,
-                sync_timeout_duration,
-                notifications_handler,
-                #[cfg(feature = "clipboard")]
-                clipboard: arboard::Clipboard::new()
-                    .inspect_err(|e| {
-                        log::warn!("couldn't create clipboard context: {e}");
-                    })
-                    .ok(),
-            })),
-        })
+            state,
+        }
     }
 
     pub async fn run(
@@ -167,7 +107,7 @@ impl Agent {
 
 async fn handle_request(
     sock: &mut crate::sock::Sock,
-    state: std::sync::Arc<tokio::sync::Mutex<State>>,
+    state: std::sync::Arc<tokio::sync::Mutex<crate::state::State>>,
 ) -> anyhow::Result<()> {
     let req = sock.recv().await?;
     let req = match req {
