+++ src/bin/rbw-agent/state.rs
@@ -1,0 +1,38 @@
+pub struct State {
+    pub priv_key: Option<rbw::locked::Keys>,
+    pub org_keys:
+        Option<std::collections::HashMap<String, rbw::locked::Keys>>,
+    pub timeout: crate::timeout::Timeout,
+    pub timeout_duration: std::time::Duration,
+    pub sync_timeout: crate::timeout::Timeout,
+    pub sync_timeout_duration: std::time::Duration,
+    pub notifications_handler: crate::notifications::Handler,
+    #[cfg(feature = "clipboard")]
+    pub clipboard: Option<arboard::Clipboard>,
+}
+
+impl State {
+    pub fn key(&self, org_id: Option<&str>) -> Option<&rbw::locked::Keys> {
+        org_id.map_or(self.priv_key.as_ref(), |id| {
+            self.org_keys.as_ref().and_then(|h| h.get(id))
+        })
+    }
+
+    pub fn needs_unlock(&self) -> bool {
+        self.priv_key.is_none() || self.org_keys.is_none()
+    }
+
+    pub fn set_timeout(&self) {
+        self.timeout.set(self.timeout_duration);
+    }
+
+    pub fn clear(&mut self) {
+        self.priv_key = None;
+        self.org_keys = None;
+        self.timeout.clear();
+    }
+
+    pub fn set_sync_timeout(&self) {
+        self.sync_timeout.set(self.sync_timeout_duration);
+    }
+}
